refactor: split AuthRepository interface from AuthImpl into separate files
